transfer the files in forgetPassword to authController
