+++ src/controllers/AuthController.ts
@@ -14,6 +14,7 @@ import emailNotVerified from '../middlewares/emailNotVerified';
 declare module 'express-session' {
   interface Session {
     verificationCode: number;
+    resetCode?: number;
   }
 }
 
@@ -80,44 +81,48 @@ router.post(
   },
 );
 
-router.post('/login', loginValidator, async (req: Request, res: Response) => {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    return res.status(400).json(formatValidationMessages(errors.array()));
-  }
-  const { email, password } = req.body;
-  const user = await User.findOne({ email });
+router.post(
+  '/login',
+  [loginValidator.emailValidator, loginValidator.passwordValidator],
+  async (req: Request, res: Response) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json(formatValidationMessages(errors.array()));
+    }
+    const { email, password } = req.body;
+    const user = await User.findOne({ email });
 
-  if (!user || !user.comparePassword(password)) {
-    return res.status(401).json({
-      message: 'Authentication failed, invalid Email or Password.',
+    if (!user || !user.comparePassword(password)) {
+      return res.status(401).json({
+        message: 'Authentication failed, invalid Email or Password.',
+      });
+    }
+
+    user.password = '';
+
+    const payload = {
+      id: user._id,
+      email: user.email,
+      userType: user.userType,
+    };
+
+    return res.status(200).json({
+      token: jwt.sign(payload, process.env.jwtSecret, { expiresIn: 15768000 }),
+      user,
+      message: user.isEmailVerified
+        ? `Welcome ${user.fullName}`
+        : `${user.fullName}, Please verify your account `,
     });
-  }
-
-  user.password = '';
-
-  const payload = {
-    id: user._id,
-    email: user.email,
-    userType: user.userType,
-  };
-
-  return res.status(200).json({
-    token: jwt.sign(payload, process.env.jwtSecret, { expiresIn: 15768000 }),
-    user,
-    message: user.isEmailVerified
-      ? `Welcome ${user.fullName}`
-      : `${user.fullName}, Please verify your account `,
-  });
-});
+  },
+);
 
 router.post(
   '/verifyEmail',
   emailNotVerified,
   async (req: Request, res: Response) => {
     if (
-      req.session.verificationCode && Number(req.body.verificationCode)
-      === Number(req.session.verificationCode)
+      req.session.verificationCode &&
+      Number(req.body.verificationCode) === Number(req.session.verificationCode)
     ) {
       const user = await User.findByIdAndUpdate(
         req.user._id,
@@ -135,6 +140,103 @@ router.post(
       res.status(200).json({ message: 'Email Verified Successfully!', user });
     } else {
       res.status(400).json({ message: 'Wrong Code Provided' });
+    }
+  },
+);
+
+router.post(
+  '/recoverPassword',
+  loginValidator.emailValidator,
+  async (req: Request, res: Response) => {
+    try {
+      const errors = validationResult(req);
+      if (!errors.isEmpty()) {
+        return res.status(400).json(formatValidationMessages(errors.array()));
+      }
+      const { email } = req.body;
+      const user = await User.findOne({ email });
+      if (!user) {
+        return res.status(404).json({ message: 'Invalid email address!' });
+      }
+
+      const resetCode = Math.floor(Math.random() * 100000);
+
+      req.session.resetCode = resetCode;
+
+      mailingService(
+        'Password Reset Code',
+        confirmEmailTemplate(resetCode),
+        email,
+      );
+
+      await user.save();
+
+      return res
+        .status(200)
+        .json({ message: 'Password reset code has been sent ' });
+    } catch (error) {
+      return res.status(500).json({ message: 'Server error', error });
+    }
+  },
+);
+
+router.post('/verifyCode', (req: Request, res: Response) => {
+  if (
+    req.session.resetCode &&
+    Number(req.body.resetCode) === Number(req.session.resetCode)
+  ) {
+    res.status(200).json({ message: 'Verified sucessfully' });
+  } else {
+    res.status(400).json({ message: 'Wrong reset code provided' });
+  }
+});
+
+router.post(
+  '/resetPassword/:code',
+  [loginValidator.emailValidator, loginValidator.passwordValidator],
+  async (req: Request, res: Response) => {
+    try {
+      const errors = validationResult(req);
+      if (!errors.isEmpty()) {
+        return res.status(400).json(formatValidationMessages(errors.array()));
+      }
+      const { email, password } = req.body;
+
+      if (Number(req.params.code) !== Number(req.session.resetCode)) {
+        return res
+          .status(400)
+          .json({ message: 'The code you enter is invalid' });
+      }
+
+      delete req.session.resetCode;
+
+      const user = await User.findOne({ email });
+
+      const salt: string = await bcrypt.genSalt(10);
+
+      user.password = await bcrypt.hash(password, salt);
+
+      await user.save();
+
+      user.password = '';
+
+      const payload = {
+        id: user._id,
+        email: user.email,
+        userType: user.userType,
+      };
+
+      return res.status(200).json({
+        token: jwt.sign(payload, process.env.jwtSecret, {
+          expiresIn: 15768000,
+        }),
+        user,
+        message: user.isEmailVerified
+          ? `Welcome ${user.fullName}`
+          : `${user.fullName}, Please verify your account `,
+      });
+    } catch (error) {
+      return res.status(500).json({ message: 'Server error', error });
     }
   },
 );
