+++ src/routes/v1.ts
@@ -1,7 +1,6 @@
 import { Router, Request, Response } from 'express';
 
 import AuthController from '../controllers/AuthController';
-import ForgotPassword from '../controllers/ForgotPassword';
 
 const router = Router();
 
@@ -10,6 +9,5 @@ router.get('/', (req: Request, res: Response) => {
 });
 
 router.use('/auth', AuthController);
-router.use('/forgotPassword', ForgotPassword);
 
 export default router;
